Make the program work for any dimension (> 2).
The program now works regardless of the dimension -- we don't have
nested loops anymore.

In the init(), we calculate the "offsets" for the given dimension,
that is, set of -1/0/1 deltas in each axis (not all 0) to jump
to the various neighbours. 3^d - 1 in total, were d is the number
of dimensions.

neigbourhood() uses the calculated offsets to generate the sets
of neighbouring cells.

This is about 2 seconds slower than the nested loop solution (which
only worked for dimensions 3 and 4). Without the caching in
neigbourhood, the program would be 3 more seconds slower.